Refactor: Separate directives
